KVM: PPC: Book3S HV: Reduce cases where H_CONFER exits guest
Currently the H_CONFER hcall is implemented in kernel virtual mode,
meaning that whenever a guest thread does an H_CONFER, all the threads
in that virtual core have to exit the guest.  This is bad for
performance because it interrupts the other threads even if they
are doing useful work.

To avoid having the other threads exit the guest unnecessarily,
we add a real-mode implementation of H_CONFER that checks whether
the other threads are doing anything.  If all the other threads
are idle (i.e. in H_CEDE), it returns H_TOO_HARD which causes a
guest exit and allows the H_CONFER to be handled by the existing
virtual-mode implementation.  Otherwise it returns H_SUCCESS,
meaning that the H_CONFER is basically a no-op.  Since H_CONFER is
called when a vcpu is spinning on a spinlock, this is OK; the
caller just continues to spin trying to get the spinlock.

BZ: 105677

Signed-off-by: Paul Mackerras <paulus@samba.org>