rtc/tpo: Driver to support rtc and wakeup on PowerNV platform
The patch implements the OPAL rtc driver that binds with the rtc
driver subsystem. The driver uses the platform device infrastructure
to probe the rtc device and register it to rtc class framework. The
'wakeup' is supported depending upon the property 'has-tpo' present
in the OF node. It provides a way to load the generic rtc driver in
in the absence of an OPAL driver.

The patch also moves the existing OPAL rtc get/set time interfaces to the
new driver and exposes the necessary OPAL calls using EXPORT_SYMBOL_GPL.

Test results:
-------------
Host:
[root@tul169p1 linus]# ls -l /sys/class/rtc/
total 0
lrwxrwxrwx 1 root root 0 Aug 28 03:21 rtc0 -> ../../devices/opal-rtc/rtc/rtc0
[root@tul169p1 linus]#
[root@tul169p1 linus]# cat /sys/devices/opal-rtc/rtc/rtc0/time
08:22:16
[root@tul169p1 linus]# cat /sys/devices/opal-rtc/rtc/rtc0/time
08:23:30
[root@tul169p1 linus]# echo `date '+%s' -d '+ 3 minutes'` > /sys/class/rtc/rtc0/wakealarm
[root@tul169p1 linus]# cat /sys/class/rtc/rtc0/wakealarm
1409214402
[root@tul169p1 linus]#

FSP:
$ smgr mfgState
standby
$ rtim timeofday

System time is valid: 2014/08/28 08:25:16.128924

$ smgr mfgState
ipling
$

Signed-off-by: Neelesh Gupta <neelegup@linux.vnet.ibm.com>
Signed-off-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>