KVM: PPC: Book3S HV: Improve H_CONFER
This improves the implementation of the H_CONFER virtual mode handler.
In the case where the target VCPU is able to run in the guest, the
responsibility for running it may be taken by another VCPU in the same
virtual core (the 'runner' VCPU).  In that case we yield to the runner
VCPU rather than the target VCPU.

BZ: 105677

Signed-off-by: Sam Bobroff <sam.bobroff@au1.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>