cpufreq/powernv: Set core pstate to a minimum just before hotplugging it out
It's possible today that the pstate of a core is held at a high even
after the entire core is hotplugged out if a load had just run on the
hotplugged cpu. This is fair, since it is assumed that the pstate does
not matter to a cpu in a deep idle state, which is the expected state
of a hotplugged core on powerpc. However on powerpc, the pstate at a
socket level is held at the maximum of the pstates of each core. Even
if the pstates of the active cores on that socket is low, the socket
pstate is held high due to the pstate of the hotplugged core in the
above mentioned scenario. This can cost significant amount of power
loss for no good.

Besides, since it is a non active core, nothing can be done from the
kernel's end to set the frequency of the core right. Hence make use of
the exit callback to explicitly set the pstate of the core to a
minimum when the last cpu of the core gets hotplugged out.

Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>