KVM: PPC: Book3S HV: Handle concurrent H_CONFERs better
Allow H_CONFER to (potentially) succeed when there is more than one
VCPU in a vcore using it concurrently.

Currently, a guest thread that is calling H_CONFER while attempting to
acquire a spinlock will appear busy to other VCPUs in its vcore.
This will prevent any of them from yielding.

We introduce a bitmask to the vcore (conferring_threads) and a short
spin loop to the H_CONFER real mode handler so that threads can see
that other, otherwise busy, VCPUs are also attempting to yield.

We also introduce a check on the target VCPUs yield count: if it
differs from the yield count passed to H_CONFER, the target VCPU
has run since H_CONFER was called and may have already released
the lock.  This check is required by PAPR.

Signed-off-by: Sam Bobroff <sam.bobroff@au1.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>