KVM: PPC: Book3S HV: Temporary fix for KSM memory corruption
Code inspection in kvmppc_h_protect revealed that there was a race
condition with the possibility of granting write access to a page
which is read-only in the host page tables. To close this race, we
don't let a page go from being read-only to being read-write, as far
as the actual mapping of the page is concerned (the guest view can go
to read-write, but the actual mapping stays read-only). When the guest
tries to write to the page, we take an HDSI and let
kvmppc_book3s_hv_page_fault take care of providing a writable HPTE for
the page.

This appears to eliminate the occasional corruption of shared pages
that was previously seen with KSM active.

BZ: 115422

Signed-off-by: Paul Mackerras <paulus@samba.org>