subcore: powernv: cpuidle: Disable fastsleep and winkle due to timebase issue
Currently we resync timebase when any thread wakes up from
fastsleep/winkle.  This resync updates timebase across all
subcores. In case where a subcore is in guest running with a different
timebase value and a thread in another subcore (but same core) wakes
up from fastsleep/winkle we end up changing the timebase for the guest
and consequently back in host when we exit the guest.  This causes
softlockup in host since timebase are out-of-sync.

Fix to this problem will need to re-syncing timebase for only the
first thread coming out of fast-sleep/winkle in the core.

This patch is a work-around to disable fast-sleep and winkle support
until a fix in implemented in the kernel.

BZ: 117526

Signed-off-by: Shreyas B. Prabhu <shreyas@linux.vnet.ibm.com>
Signed-off-by: Vaidyanathan Srinivasan <svaidy@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>