KVM: PPC: Book3S HV: Make sure page fault puts HPTE in correct revmap chain
When we add a large page HPTE, it is critical to ensure that it gets
added to the rmap chain for its base address, so that
remove_revmap_chain() locks the correct rmap chain when removing it.
Currently in kvmppc_book3s_hv_page_fault() we choose the rmap chain
for the fault address, which for huge pages may not be the rmap chain
for the page base address.  This fixes it by using the page base gfn
to select the rmap chain.

BZ: 116703

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>