powerpc: Fix nap-mode code for little-endian
The code to turn off the MMU when entering nap/sleep/rvwinkle mode
always reset the MSR to big-endian mode, causing illegal instruction
traps when the CPU tried to execute the MMU-off code in a little-
endian kernel.  This fixes it by incorporating the MSR_LE bit value
from MSR_KERNEL (which includes MSR_LE on LE kernels but not BE
kernels).

Signed-off-by: Paul Mackerras <paulus@samba.org>