KVM: Move vcpu_id range check to arch code
Commit f4d838225e0c ("KVM: Improve create VCPU parameter
(CVE-2013-4587)") added a check that the VCPU id supplied to the
KVM_CREATE_VCPU ioctl is less than KVM_MAX_VCPUS, because some
code uses the id as a bit number in accessing a bitmap.  In fact
only ARM, X86 and IA64 do that.

On powerpc machines using the book3s_hv code, the added check
restricts the number of vcpus we can create in a VM to only
NR_CPUS/8 if the vcpus are single-threaded, because of the way
that vcpu IDs are assigned.

This restores the previous behavior on powerpc book3s_hv by moving
the added check into the architectures that need it.  In fact,
ia64 and s390 already had the check.

Signed-off-by: Paul Mackerras <paulus@samba.org>