KVM: PPC: Book3S HV: Use vcore pointer rather than vcpu to trigger secondaries
This is in preparation for the next patch, which will implement dynamic
split-core mode, where we will have secondary threads which are the
primary thread for their subcore and therefore need to do the MMU
switch.  These threads will need to be started even if they have no
vcpu to run, so this provides a way for them to be started.

Signed-off-by: Paul Mackerras <paulus@samba.org>